Fixes after merging
