+++ src/client/datascience/jupyter/kernels/kernelSelections.ts
@@ -4,16 +4,21 @@
 'use strict';
 
 import { inject, injectable } from 'inversify';
-import { CancellationToken, QuickPickItem } from 'vscode';
+import { CancellationToken } from 'vscode';
 import { IInterpreterSelector } from '../../../interpreter/configuration/types';
-import { PythonInterpreter } from '../../../interpreter/contracts';
 import { IJupyterKernelSpec, IJupyterSessionManager } from '../../types';
 import { KernelService } from './kernelService';
+import { IKernelSelectionListProvider, IKernelSpecQuickPickItem } from './types';
 
-export interface IKernelSpecQuickPickItem extends QuickPickItem {
-    selection: { kernelSpec: IJupyterKernelSpec } | { interpreter: PythonInterpreter };
-}
+// Small classes, hence all put into one file.
+// tslint:disable: max-classes-per-file
 
+/**
+ * Given a kernel spec, this will return a quick pick item with appropriate display names and the like.
+ *
+ * @param {IJupyterKernelSpec} kernelSpec
+ * @returns {IKernelSpecQuickPickItem}
+ */
 function getQuickPickFromKernelSpec(kernelSpec: IJupyterKernelSpec): IKernelSpecQuickPickItem {
     return {
         label: kernelSpec.display_name || kernelSpec.name || '',
@@ -21,53 +26,15 @@ function getQuickPickFromKernelSpec(kernelSpec: IJupyterKernelSpec): IKernelSpec
         selection: { kernelSpec: kernelSpec }
     };
 }
-export interface IKernelSelector {
-    selectKernelSpec(cancelToken?: CancellationToken): Promise<IJupyterKernelSpec | undefined>;
-}
 
-// export class kernelSelector implements IKernelSelector {
-//     public async selectKernelSpec(session?: IJupyterSessionManager, cancelToken?: CancellationToken): Promise<IJupyterKernelSpec | undefined> {
-//         return;
-//     }
-//     private selectKernelSpecFromList(session?: IJupyterSessionManager, cancelToken?: CancellationToken): Promise<IJupyterKernelSpec | undefined> {}
-// }
-
-export interface IKernelSelectionListProvider {
-    getKernelSelections(cancelToken?: CancellationToken): Promise<IKernelSpecQuickPickItem[]>;
-}
-
-
-@injectable()
-export class KernelSelectionProviderFactory {
-    constructor(@inject(KernelService) private readonly kernelService: KernelService, @inject(IInterpreterSelector) private readonly interpreterSelector: IInterpreterSelector) {}
-    public async getRemoteKernelSelectionProvider(session: IJupyterSessionManager): Promise<IKernelSelectionListProvider> {
-        // tslint:disable: no-use-before-declare
-        return new RemoteKernelSelectionProvider(session);
-    }
-    public async getLocalKernelSelectionProvider(session?: IJupyterSessionManager): Promise<IKernelSelectionListProvider> {
-        return new LocalKernelSelectionProvider(this.kernelService, this.interpreterSelector, session);
-    }
-}
-export class RemoteKernelSelectionProvider implements IKernelSelectionListProvider {
-    constructor(private readonly session: IJupyterSessionManager) {}
-    public async getKernelSelections(cancelToken?: CancellationToken): Promise<IKernelSpecQuickPickItem[]> {
-        return new JupyterSessionKernelSelectionnListProvider(this.session).getKernelSelections(cancelToken);
-    }
-}
-
-// tslint:disable: max-classes-per-file
-export class LocalKernelSelectionProvider implements IKernelSelectionListProvider {
-    constructor(private readonly kernelService: KernelService, private readonly interpreterSelector: IInterpreterSelector, private readonly session?: IJupyterSessionManager) {}
-    public async getKernelSelections(cancelToken?: CancellationToken): Promise<IKernelSpecQuickPickItem[]> {
-        const activeKernelsPromise = this.session ? new JupyterSessionKernelSelectionnListProvider(this.session).getKernelSelections(cancelToken) : Promise.resolve([]);
-        const jupyterKernelsPromise = new JupyterKernelSelectionnListProvider(this.kernelService).getKernelSelections(cancelToken);
-        const interpretersPromise = new InterpreterKernelSelectionnListProvider(this.interpreterSelector).getKernelSelections(cancelToken);
-        const [activeKernels, jupyterKernels, interprters] = await Promise.all([activeKernelsPromise, jupyterKernelsPromise, interpretersPromise]);
-        return [...activeKernels!, ...jupyterKernels!, ...interprters];
-    }
-}
-
-export class JupyterSessionKernelSelectionnListProvider implements IKernelSelectionListProvider {
+/**
+ * Provider for active kernel specs in a jupyter session.
+ *
+ * @export
+ * @class ActiveJupyterSessionKernelSelectionListProvider
+ * @implements {IKernelSelectionListProvider}
+ */
+export class ActiveJupyterSessionKernelSelectionListProvider implements IKernelSelectionListProvider {
     constructor(private readonly session: IJupyterSessionManager) {}
     public async getKernelSelections(_cancelToken?: CancellationToken | undefined): Promise<IKernelSpecQuickPickItem[]> {
         const items = await this.session.getActiveKernelSpecs();
@@ -75,6 +42,13 @@ export class JupyterSessionKernelSelectionnListProvider implements IKernelSelect
     }
 }
 
+/**
+ * Provider for kernel specs in a jupyter process (`python -m jupyter kernelspec list`).
+ *
+ * @export
+ * @class JupyterKernelSelectionnListProvider
+ * @implements {IKernelSelectionListProvider}
+ */
 export class JupyterKernelSelectionnListProvider implements IKernelSelectionListProvider {
     constructor(private readonly kernelService: KernelService) {}
     public async getKernelSelections(cancelToken?: CancellationToken | undefined): Promise<IKernelSpecQuickPickItem[]> {
@@ -83,7 +57,13 @@ export class JupyterKernelSelectionnListProvider implements IKernelSelectionList
     }
 }
 
-// tslint:disable: max-classes-per-file
+/**
+ * Provider for interpreters to be treated as kernel specs.
+ *
+ * @export
+ * @class InterpreterKernelSelectionnListProvider
+ * @implements {IKernelSelectionListProvider}
+ */
 export class InterpreterKernelSelectionnListProvider implements IKernelSelectionListProvider {
     constructor(private readonly interpreterSelector: IInterpreterSelector) {}
     public async getKernelSelections(_cancelToken?: CancellationToken | undefined): Promise<IKernelSpecQuickPickItem[]> {
@@ -94,5 +74,42 @@ export class InterpreterKernelSelectionnListProvider implements IKernelSelection
                 selection: { interpreter: item.interpreter }
             };
         });
+    }
+}
+
+/**
+ * Factory class that provides a kernel spec list provider (local or remote).
+ *
+ * @export
+ * @class KernelSelectionProviderFactory
+ */
+@injectable()
+export class KernelSelectionProvider {
+    constructor(@inject(KernelService) private readonly kernelService: KernelService, @inject(IInterpreterSelector) private readonly interpreterSelector: IInterpreterSelector) {}
+    /**
+     * Gets a selection of kernel specs from a remote session.
+     *
+     * @param {IJupyterSessionManager} session
+     * @param {CancellationToken} [cancelToken]
+     * @returns {Promise<IKernelSpecQuickPickItem[]>}
+     * @memberof KernelSelectionProvider
+     */
+    public async getKernelSelectionsForRemoteSession(session: IJupyterSessionManager, cancelToken?: CancellationToken): Promise<IKernelSpecQuickPickItem[]> {
+        return new ActiveJupyterSessionKernelSelectionListProvider(session).getKernelSelections(cancelToken);
+    }
+    /**
+     * Gets a selection of kernel specs for a local session.
+     *
+     * @param {IJupyterSessionManager} [session]
+     * @param {CancellationToken} [cancelToken]
+     * @returns {Promise<IKernelSelectionListProvider>}
+     * @memberof KernelSelectionProvider
+     */
+    public async getLocalKernelSelectionProvider(session?: IJupyterSessionManager, cancelToken?: CancellationToken): Promise<IKernelSpecQuickPickItem[]> {
+        const activeKernelsPromise = session ? new ActiveJupyterSessionKernelSelectionListProvider(session).getKernelSelections(cancelToken) : Promise.resolve([]);
+        const jupyterKernelsPromise = new JupyterKernelSelectionnListProvider(this.kernelService).getKernelSelections(cancelToken);
+        const interpretersPromise = new InterpreterKernelSelectionnListProvider(this.interpreterSelector).getKernelSelections(cancelToken);
+        const [activeKernels, jupyterKernels, interprters] = await Promise.all([activeKernelsPromise, jupyterKernelsPromise, interpretersPromise]);
+        return [...activeKernels!, ...jupyterKernels!, ...interprters];
     }
 }
